refactor: put types in global file
